+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/actuators/actuator_cfg.py
@@ -10,8 +10,8 @@ from typing import Iterable
 
 from omni.isaac.orbit.utils import configclass
 
-from .actuator import ActuatorBase, DCMotor, IdealPDActuator, ImplicitActuator
-from .actuator_net import ActuatorNetLSTM, ActuatorNetMLP
+from . import actuator_net, actuator_pd
+from .actuator_base import ActuatorBase
 
 
 @configclass
@@ -21,56 +21,71 @@ class ActuatorBaseCfg:
     cls: type[ActuatorBase] = MISSING
     """Actuator class."""
 
-    dof_name_expr: list[str] = MISSING
-    """Articulation's DOF names that are part of the group. Can be regex expressions (e.g. ".*")."""
+    dof_names_expr: list[str] = MISSING
+    """Articulation's joint names that are part of the group.
+
+    Note:
+        This can be a list of joint names or a list of regex expressions (e.g. ".*").
+    """
 
     effort_limit: float | None = None
-    """
-    Force/Torque limit of the DOFs in the group. Defaults to :obj:`None`.
+    """Force/Torque limit of the joints in the group. Defaults to :obj:`None`.
+
+    If :obj:`None`, the limit is set to infinity.
     """
 
     velocity_limit: float | None = None
-    """
-    Velocity limit of the DOFs in the group. Defaults to :obj:`None`.
+    """Velocity limit of the joints in the group. Defaults to :obj:`None`.
+
+    If :obj:`None`, the limit is set to infinity.
     """
 
-    stiffness: dict[str, float] | None = None
-    """
-    Stiffness gains (Also known as P gain) of the DOFs in the group. Defaults to :obj:`None`.
-    If :obj:`None`, these are loaded from the articulation prim.
+    stiffness: dict[str, float] | None = MISSING
+    """Stiffness gains (also known as p-gain) of the joints in the group.
+
+    If :obj:`None`, the stiffness is set to 0.
     """
 
-    damping: dict[str, float] | None = None
+    damping: dict[str, float] | None = MISSING
+    """Damping gains (also known as d-gain) of the joints in the group.
+
+    If :obj:`None`, the damping is set to 0.
     """
-    Damping gains (Also known as D gain) of the DOFs in the group. Defaults to :obj:`None`.
-    If :obj:`None`, these are loaded from the articulation prim.
+
+
+"""
+Implicit Actuator Models.
+"""
+
+
+@configclass
+class ImplicitActuatorCfg(ActuatorBaseCfg):
+    """Configuration for an ideal PD actuator.
+
+    Note:
+        The PD control is handled implicitly by the simulation.
     """
+
+    cls = actuator_pd.ImplicitActuator
+
+
+"""
+Explicit Actuator Models.
+"""
 
 
 @configclass
 class IdealPDActuatorCfg(ActuatorBaseCfg):
-    """Configuration for an ideal PD actuator.
-    The PD control is handled implicitly by the simulation."""
+    """Configuration for an ideal PD actuator."""
 
-    cls: type[ActuatorBase] = IdealPDActuator
-    """Actuator class."""
-
-
-@configclass
-class ImplicitPDActuatorCfg(ActuatorBaseCfg):
-    """Configuration for an ideal PD actuator.
-    The PD control is handled implicitly by the simulation."""
-
-    cls: type[ActuatorBase] = ImplicitActuator
-    """Actuator class."""
+    cls = actuator_pd.IdealPDActuator
 
 
 @configclass
 class DCMotorCfg(IdealPDActuatorCfg):
     """Configuration for direct control (DC) motor actuator model."""
 
-    cls: type[ActuatorBase] = DCMotor
-    """Actuator class."""
+    cls = actuator_pd.DCMotor
 
     saturation_effort: float = MISSING
     """Peak motor force/torque of the electric DC motor (in N-m)."""
@@ -80,7 +95,10 @@ class DCMotorCfg(IdealPDActuatorCfg):
 class ActuatorNetLSTMCfg(DCMotorCfg):
     """Configuration for LSTM-based actuator model."""
 
-    cls: type[ActuatorBase] = ActuatorNetLSTM
+    cls = actuator_net.ActuatorNetLSTM
+    # we don't use stiffness and damping for actuator net
+    stiffness = None
+    damping = None
 
     network_file: str = MISSING
     """Path to the file containing network weights."""
@@ -90,7 +108,10 @@ class ActuatorNetLSTMCfg(DCMotorCfg):
 class ActuatorNetMLPCfg(DCMotorCfg):
     """Configuration for MLP-based actuator model."""
 
-    cls: type[ActuatorBase] = ActuatorNetMLP
+    cls = actuator_net.ActuatorNetMLP
+    # we don't use stiffness and damping for actuator net
+    stiffness = None
+    damping = None
 
     network_file: str = MISSING
     """Path to the file containing network weights."""