+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/actuators/config/anydrive.py
@@ -10,9 +10,6 @@ The following actuator models are available:
 * ANYdrive 3.x with DC actuator model.
 * ANYdrive 3.0 (used on ANYmal-C) with LSTM actuator model.
 
-The following actuator groups are available:
-
-* ANYmal-C default actuator group that uses ANYdrive 3.0 with LSTM actuator model.
 """
 
 from omni.isaac.orbit.actuators import ActuatorNetLSTMCfg, DCMotorCfg